requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0
sse-starlette>=1.8.0
jinja2>=3.1.2

# 文档解析
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
PyMuPDF>=1.23.0
chardet>=5.2.0

# 数据处理
pydantic>=2.5.0
pyyaml>=6.0.1
